crypto/tls: adjust boring_test.go for the recent changes in 1.21
In Go 1.21 upstream, 4096-bit RSA keys are supported in the Boring
mode[1][2], which causes key generation failure when backed with
OpenSSL.  This instead uses 2560-bit RSA key to cover the scenario,
which is supported by OpenSSL but not by the Boring mode.

This also adds a check on SHA1 signature usages in TLS and skips them
if our test config doesn't support it.

1. https://github.com/golang/go/commit/7bc3281747030877e13d218ba12c6e95fcf4e7d4
2. https://github.com/golang/go/commit/9ff2a6677fd34084295172aff0848ab9a4f5b76e

Signed-off-by: Daiki Ueno <dueno@redhat.com>